ListViewRepo ReadMe
This is a sample app which shows a smooth way to replace the element (say a layout) in a ListView with another element (a different layout) when an element of the listView is clicked. This app only shows how to switch the views (for now), and can be extended to have a more appealing UI by the use of appropriate animations.